backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# SAML Authentication
python3-saml==1.16.0

# Session Management
Flask-Session==0.5.0
redis==5.0.1

# CORS Support
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Meraki Dashboard API (or: pip install -e /path/to/dashboard-api-python)
meraki

# Optional: Database support (uncomment if needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# SQLAlchemy==2.0.23

# Development tools
pytest==7.4.3
pytest-flask==1.3.0

